pyproject.toml
[project]
name = "pyobs-tui"
version = "0.0.1"
description = "A remote TUI for pyobs"
authors = [{ name = "Tim-Oliver Husser", email = "thusser@uni-goettingen.de" }]
requires-python = ">=3.11,<=3.14"
license = "MIT"
dependencies = [
    "pyobs-core>=1.24.0",
    "textual>=6.7.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0,<26",
    "flake8>=7.3.0",
    "pre-commit>=4.2.0,<5",
]

[build-system]
requires = ["uv_build>=0.9.14,<0.10.0"]
build-backend = "uv_build"

[tool.black]
line-length = 120
target-version = ['py311']
